fast_react/requirements.txt
aiosmtplib==3.0.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
fastapi-mail==1.5.0
h11==0.16.0
idna==3.10
iso8601==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pypika-tortoise==0.6.1
python-dotenv==1.1.1
pytz==2025.2
sniffio==1.3.1
starlette==0.47.2
tortoise-orm==0.25.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
